Create concatlinkedlist.c++
Concatenation of Two Singly Linked Lists in C++:

Concatenation of two singly linked lists involves combining the elements of one list with the elements of another, creating a longer linked list. This operation can be done by changing the next pointer of the last node of the first list to point to the head of the second list. It's a straightforward operation that allows you to merge two linked lists into a single list, preserving their individual order.